# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Kanye West Quotes",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the Bearer Token",
                "description": "Return bearer token",
                "operationId": "getBearerToken",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content Type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "description": "Request body is required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "email": "test@test.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful token generated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/quotes/": {
            "get": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Get list of your favorite Quotes",
                "description": "Returns a list of your favorite Kanye West Quotes",
                "operationId": "getListOfQuotes",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content Type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Quotes objects based on your favorites",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/random_quotes/{amount}": {
            "get": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Get list of Random Quotes",
                "description": "Returns list of random generated quotes",
                "operationId": "getListOfRandomQuotes",
                "parameters": [
                    {
                        "name": "amount",
                        "in": "path",
                        "description": "The number of random quotes to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content Type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of random generated quotes",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/quote/{quoteId}": {
            "delete": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Delete a Quote from your favorites.",
                "description": "Delete a Quote from your favorites list of Kanye West Quotes",
                "operationId": "deleteQuoteFromFavorites",
                "parameters": [
                    {
                        "name": "quoteId",
                        "in": "path",
                        "description": "The id of the Quote to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content Type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the deleted object",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}